+++ tests/covalent_ui_backend_tests/utils/data/mock_files.py
@@ -16,6 +16,7 @@
 
 """Mock files data"""
 
+import json
 import os
 import pickle
 
@@ -50,12 +51,12 @@ def mock_files_data():
         "function": "function.pkl",
         "executor": "executor_data.pkl",
         "deps": "deps.pkl",
-        "cova_imports": "cova_imports.pkl",
+        "cova_imports": "cova_imports.json",
         "error": "error.log",
         "function_docstring": "function_docstring.txt",
         "function_string": "function_string.txt",
         "inputs": "inputs.pkl",
-        "lattice_imports": "lattice_imports.pkl",
+        "lattice_imports": "lattice_imports.txt",
     }
 
     _object_id = "gAWVNwAAAAAAAACMG2NvdmFsZW50Ll93b3JrZmxvdy5kZXBzYmFzaJSME2FwcGx5X2Jhc2hfY29tbWFuZHOUk5Qu"  # pragma: allowlist secret
@@ -78,7 +79,7 @@ def mock_files_data():
             "files": [
                 {"file_name": file_name["call_after"], "data": []},
                 {"file_name": file_name["call_before"], "data": []},
-                {"file_name": file_name["cova_imports"], "data": {"electron", "ct"}},
+                {"file_name": file_name["cova_imports"], "data": json.dumps(["electron", "ct"])},
                 {"file_name": file_name["deps"], "data": {}},
                 {"file_name": file_name["error"], "data": ""},
                 {"file_name": file_name["executor"], "data": {}},